feat: move domain.Expression to evaluator package
