+++ evaluator/expression.go
@@ -1,4 +1,4 @@
-package domain
+package evaluator
 
 import (
 	"errors"
